Client can now download files from a server. Tested it and successfully downloaded a picture of Aaron. Client is now better formatted.

+++ EZShare/src/EZShare/Resource.java
@@ -21,7 +21,7 @@ public class Resource extends JsonModel {
 	public String channel;
 	public String owner;
 	public String ezserver;
-	public Long resourceSize;
+	public int resourceSize;
 
 	/**
 	 * Default constructor
